WIP: ré-organiser les boutons et modales d'action sur la consultation d'un dossier
