+++ assets/apps/agent/dossiers/components/consultation/action/index.ts
@@ -1,0 +1,3 @@
+import { DossierActions } from "./DossierActions";
+
+export { DossierActions };
