+++ assets/apps/agent/dossiers/components/consultation/editor/index.ts
@@ -1,0 +1,3 @@
+import { QuillEditor } from "./QuillEditor";
+
+export { QuillEditor };
