--- assets/apps/agent/dossiers/components/consultation/index.ts
@@ -1,13 +1,0 @@
-import { AttributionDossier } from "./AttributionDossier";
-import { ClotureDossier } from "./ClotureDossier";
-import { DecisionDossier } from "./DecisionDossier";
-import { ValidationDecisionDossier } from "./ValidationDecisionDossier";
-import { ValidationAcceptationDossier } from "./ValidationAcceptationDossier";
-
-export {
-  AttributionDossier,
-  ClotureDossier,
-  DecisionDossier,
-  ValidationDecisionDossier,
-  ValidationAcceptationDossier,
-};
