+++ assets/apps/agent/dossiers/components/consultation/piecejointe/index.ts
@@ -1,0 +1,4 @@
+import { MetaDonneesAttestationForm } from "./MetaDonneesAttestation";
+import { PieceJointe } from "./PieceJointe";
+
+export { MetaDonneesAttestationForm, PieceJointe };
