+++ assets/apps/agent/dossiers/models/Dossier.ts
@@ -1,18 +1,8 @@
-import { EtatDossier } from "@/apps/agent/dossiers/models/EtatDossier";
-import { InstitutionSecuritePublique } from "@/apps/agent/dossiers/models/InstitutionSecuritePublique";
-import { object } from "prop-types";
-import {
-  Adresse,
-  Courrier,
-  Document,
-  DocumentType,
-  EtatDossierType,
-  Redacteur,
-  Requerant,
-  TestEligibilite,
-} from ".";
-import { Expose, plainToInstance, Transform, Type } from "class-transformer";
-import { action, computed, makeObservable, observable } from "mobx";
+import {EtatDossier} from "@/apps/agent/dossiers/models/EtatDossier";
+import {InstitutionSecuritePublique} from "@/apps/agent/dossiers/models/InstitutionSecuritePublique";
+import {Adresse, Courrier, Document, DocumentType, EtatDossierType, Redacteur, Requerant, TestEligibilite,} from ".";
+import {Expose, plainToInstance, Transform, Type} from "class-transformer";
+import {action, computed, makeObservable, observable} from "mobx";
 
 export abstract class BaseDossier {
   public readonly id: number;
@@ -210,5 +200,9 @@ export class DossierDetail extends BaseDossier {
 
   public viderDocumentParType(type: DocumentType): void {
     this.documents.set(type.type, []);
+  }
+
+  public getCourrierAJour(): Document | Courrier {
+    return this.documents.get(DocumentType.TYPE_COURRIER_REQUERANT.type)?.at(0) ?? this.documents.get(DocumentType.TYPE_COURRIER_MINISTERE.type)?.at(0) ?? this.courrier
   }
 }
